third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-selection-preservation-expected.txt

This tests for problems where we'd lose the selection in a textarea when making style and value changes.

- default value
PASS ta.selectionStart is 0
PASS ta.selectionEnd is 0
- set selectionStart/End
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- add background style
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- set value to same value
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- set value to a different value
PASS ta.selectionStart is 6
PASS ta.selectionEnd is 6
- set selection so we can test again without a trailing newline
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- add background style
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- set value to same value
PASS ta.selectionStart is 3
PASS ta.selectionEnd is 4
- reset form
PASS ta.selectionStart is 7
PASS ta.selectionEnd is 7
- set new defaultValue
PASS ta.selectionStart is 9
PASS ta.selectionEnd is 9
- set same defaultValue
PASS ta.selectionStart is 9
PASS ta.selectionEnd is 9
- append a text node
PASS ta.selectionStart is 12
PASS ta.selectionEnd is 12
- append a empty text node
PASS ta.selectionStart is 2
PASS ta.selectionEnd is 3
PASS successfullyParsed is true

TEST COMPLETE
